+++ cheetah/utils/__init__.py
@@ -1,0 +1,2 @@
+from .kde import kde_histogram_1d, kde_histogram_2d  # noqa: F401
+from .unique_name_generator import UniqueNameGenerator  # noqa: F401
