+++ cheetah/utils/unique_name_generator.py
@@ -1,0 +1,11 @@
+class UniqueNameGenerator:
+    """Generates a unique name given a prefix."""
+
+    def __init__(self, prefix: str):
+        self._prefix = prefix
+        self._counter = 0
+
+    def __call__(self):
+        name = f"{self._prefix}_{self._counter}"
+        self._counter += 1
+        return name
